feat(train script): add auto logging

+++ src/train.py
@@ -16,6 +16,7 @@ import os
 import pandas as pd
 import mlflow
 import mlflow.keras
+import mlflow.tensorflow
 import shutil
 
 from pathlib import Path
@@ -63,22 +64,31 @@ print("[ output model ]\n",output_model)
 with mlflow.start_run() as run:
     experiment_artifact_path=os.path.join(artifact_path ,"training") ####
     model_artifact_path=os.path.join(experiment_artifact_path,"model_artifacts")
+    print("----mlflow.get_artifact_uri() : ",mlflow.get_artifact_uri())
 
     
-    history=model_.fit(generatorobjet.train_generator,
+    """history=model_.fit(generatorobjet.train_generator,
                    epochs= params.nbr_epoch,
-                   validation_data=generatorobjet.validation_generator)
-                 
-    model_.save(os.path.join(output_model))
-    mlflow.log_params(params_dict)
-    shutil.rmtree(model_artifact_path) # remove  model_artifact_path folder 
-    mlflow.keras.save_model(model_, model_artifact_path)
-    
-    print("----mlflow.get_artifact_uri() : ",mlflow.get_artifact_uri())
-    history_df = pd.DataFrame(history.history)
+                   validation_data=generatorobjet.validation_generator)"""
+    #history_df = pd.DataFrame(history.history)
+    history_df = pd.DataFrame(["test"])
     history_df.to_csv(os.path.join(experiment_artifact_path,"history.csv" ),index=False)
-    mlflow.log_artifacts(artifact_path)
-    mlflow.keras.autolog(registered_model_name="autolog")
+                 
+    #model_.save(os.path.join(output_model))
+
+    # temporary
+    from utils.Loadingmodel_data import modelLoad,historyLoad 
+    model_=modelLoad("./../../results/training/model2.h5")
+    #----
+    if Path(model_artifact_path).exists():
+        shutil.rmtree(model_artifact_path) # remove  model_artifact_path folder 
+
+    mlflow.log_params(params_dict)
+   # mlflow.tensorflow.autolog(log_models=False)
+    mlflow.tensorflow.autolog()
+    #mlflow.keras.save_model(model_, model_artifact_path)
+    #mlflow.log_artifacts(artifact_path,artifact_path="/")
+    mlflow.keras.log_model(model_,"keras_plant_disease")
